LearningSpace 2 menu is created. LearningSpace2 stage is splitted into different parts(members,resources, etc.)
