+++ android/app/src/main/java/com/example/myapplication/view/LearningSpace2Members.kt
@@ -1,0 +1,12 @@
+package com.example.myapplication.view
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.example.myapplication.R
+
+class LearningSpace2Members : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_learning_space2_members)
+    }
+}
