+++ android/app/src/main/java/com/example/myapplication/view/LsSearchActivity.kt
@@ -94,7 +94,7 @@ class LsSearchActivity : AppCompatActivity() {
 
     }
     fun toLearningSpace2() {
-        var intent= Intent(applicationContext, LearningSpace2::class.java)
+        var intent= Intent(applicationContext, LearningSpace2Menu::class.java)
         startActivity(intent)
     }
 }
