+++ android/app/src/main/res/drawable/background_general.xml
@@ -1,6 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android" android:shape="rectangle">
-    <gradient android:type="linear" android:angle="270" android:startColor="@color/colorPrimary1"
-        android:centerColor="@color/colorPrimary2"
-        android:endColor="@color/colorPrimary4"/>
-</shape>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item>
+        <shape android:shape="rectangle">
+            <solid android:color="@color/colorPrimary4"/>
+            <corners android:radius="50dp"/>
+            <stroke android:color="@color/colorPrimary1" android:width="1dp"></stroke>
+        </shape>
+    </item>
+</selector>
