+++ android/app/src/main/res/drawable/button_background.xml
@@ -4,7 +4,7 @@
     <shape android:shape="rectangle">
         <solid android:color="@color/colorPrimary4"/>
         <corners android:radius="50dp"/>
-        <stroke android:color="@color/colorPrimary3" android:width="1dp"></stroke>
+        <stroke android:color="@color/colorPrimary1" android:width="1dp"></stroke>
     </shape>
 </item>
 </selector>
