--- android/app/src/main/res/layout/activity_learning_space2.xml
@@ -1,103 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@drawable/background_general">
-
-
-    <TextView
-        android:id="@+id/learning_topic"
-        android:layout_width="183dp"
-        android:layout_height="44dp"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="5dp"
-        android:fontFamily="@font/walter_turncoat"
-        android:text="Gitar"
-        android:textAlignment="center"
-        android:textColor="@color/colorPrimary4"
-        android:textSize="24sp"
-        app:layout_constraintBottom_toTopOf="@+id/Owners"
-        app:layout_constraintEnd_toStartOf="@+id/textView8"
-        app:layout_constraintHorizontal_bias="0.133"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.846" />
-
-    <ListView
-        android:id="@+id/Topics"
-        android:layout_width="190dp"
-        android:layout_height="527dp"
-        android:background="@drawable/textview_background"
-        app:layout_constraintBottom_toBottomOf="@+id/Owners"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.072"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="@+id/Owners"
-        app:layout_constraintVertical_bias="0.0" />
-
-    <ListView
-        android:id="@+id/Owners"
-        android:layout_width="179dp"
-        android:layout_height="387dp"
-        android:layout_marginStart="12dp"
-        android:layout_marginTop="5dp"
-        android:layout_marginEnd="7dp"
-        android:background="@drawable/textview_background"
-        app:layout_constraintBottom_toTopOf="@+id/add_content"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.73"
-        app:layout_constraintStart_toEndOf="@+id/Topics"
-        app:layout_constraintTop_toBottomOf="@+id/textView8"
-        app:layout_constraintVertical_bias="0.0" />
-
-    <TextView
-        android:id="@+id/textView8"
-        android:layout_width="174dp"
-        android:layout_height="40dp"
-        android:layout_marginTop="16dp"
-        android:fontFamily="@font/walter_turncoat"
-        android:text="Members"
-        android:textAlignment="center"
-        android:textColor="@color/colorPrimary4"
-        android:textSize="24sp"
-        app:layout_constraintBottom_toTopOf="@+id/add_content"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.925"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.012" />
-
-    <Button
-        android:id="@+id/join_leave"
-        android:layout_width="125dp"
-        android:layout_height="51dp"
-        android:background="@drawable/button_background"
-        android:fontFamily="@font/walter_turncoat"
-        android:onClick="enroll"
-        android:text="Enroll"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.903"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/add_content"
-        app:layout_constraintVertical_bias="0.228" />
-
-    <Button
-        android:fontFamily="@font/walter_turncoat"
-        android:id="@+id/add_content"
-        android:layout_width="125dp"
-        android:layout_height="55dp"
-        android:layout_marginStart="25dp"
-        android:layout_marginEnd="30dp"
-        android:layout_marginBottom="84dp"
-        android:background="@drawable/button_background"
-        android:onClick="addContent"
-        android:text="Add Content"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="1.0"
-        app:layout_constraintStart_toEndOf="@+id/Topics" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
