+++ android/app/src/main/res/layout/activity_learning_space2_seecontents.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="@drawable/background_general">
+
+
+    <TextView
+        android:id="@+id/learning_topic"
+        android:layout_width="428dp"
+        android:layout_height="56dp"
+        android:layout_marginStart="4dp"
+        android:layout_marginTop="4dp"
+        android:fontFamily="@font/walter_turncoat"
+        android:text="Name of Learning Space"
+        android:textAlignment="center"
+        android:textColor="@color/colorPrimary1"
+        android:textSize="24sp"
+        app:layout_constraintBottom_toTopOf="@+id/Topics"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <ListView
+        android:id="@+id/Topics"
+        android:layout_width="408dp"
+        android:layout_height="621dp"
+        android:layout_marginBottom="54dp"
+        android:background="@drawable/textview_background"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/learning_topic" />
+
+</androidx.constraintlayout.widget.ConstraintLayout>
