+++ android/app/src/main/res/layout/activity_sign_in.xml
@@ -123,13 +123,12 @@
             android:layout_height="63dp"
             android:ems="10"
             android:fontFamily="@font/walter_turncoat"
-            android:shadowColor="@color/colorPrimary4"
             android:shadowDx="0.0"
             android:shadowDy="0.0"
             android:shadowRadius="8"
             android:text="Budemi Login"
             android:textAlignment="center"
-            android:textColor="#cf1d1d"
+            android:textColor="@color/colorPrimary1"
             android:textFontWeight="5"
             android:textIsSelectable="false"
             android:textSize="40dp"
