+++ android/app/src/main/res/values/styles.xml
@@ -2,10 +2,12 @@
     <!-- Base application theme. -->
     <style name="AppTheme" parent="Theme.MaterialComponents.Light.Bridge">
         <!-- Customize your theme here. -->
+
         <item name="android:fontFamily">@font/walter_turncoat</item>
         <item name="colorPrimary">@color/colorPrimary1</item>
         <item name="colorPrimaryDark">@color/colorPrimary1</item>
         <item name="colorAccent">@color/colorAccent</item>
+
 
     </style>
